refactor: 커스텀 훅 만들기
