+++ src/common/components/Modal/hooks/useOutsideClick.ts
@@ -1,0 +1,16 @@
+import { RefObject, useEffect } from "react";
+
+export default function useOutSideClick(ref: RefObject<HTMLDivElement>, onClick: () => void) {
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (ref.current && !ref.current.contains(event.target as Node)) {
+        onClick();
+      }
+    };
+
+    document.addEventListener("mousedown", handleClickOutside);
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+    };
+  }, [ref, onClick]);
+}
