+++ src/pages/index.tsx
@@ -3,7 +3,7 @@ import styles from "@/pages/index.module.scss";
 import { useState } from "react";
 
 export default function Home() {
-  const [isModalOpen, setIsModalOpen] = useState(true);
+  const [isModalOpen, setIsModalOpen] = useState(false);
 
   const handleClickConfirm = () => {
     setIsModalOpen((prev) => !prev);
@@ -31,12 +31,12 @@ export default function Home() {
               leftButtonText="아니오"
               rightButtonText="취소"
             />
-            <Modal.Confirm message="등록이 완료되었습니다." className={styles.test} onClick={handleClickConfirm} />
+            {/* <Modal.Confirm message="등록이 완료되었습니다." className={styles.test} onClick={handleClickConfirm} />
             <Modal.Warn
               message="내 프로필을 먼저 등록해 주세요."
               onClick={handleClickConfirm}
               className={styles.test}
-            />
+            /> */}
           </Modal>
         </div>
       )}
